Photoshoplift/requirements.txt
Pillow
opencv-python
tk
screeninfo
numpy
matplotlib
scipy